+++ packages/flutter/android/src/main/kotlin/com/uulab/securekeypad/SecureKeypadView.kt
@@ -254,7 +254,10 @@ public open class SecureKeypadView @JvmOverloads constructor(
 
     override fun onAttachedToWindow() {
         super.onAttachedToWindow()
-        requireSecureWindow()
+        if (!ensureSecureWindowProtection()) {
+            failClosedSecureWindowBoundary()
+            return
+        }
         requestSessionReconfigurationIfNeeded()
     }